Reorganize files to be uniformly structured
