+++ src/renderer/jobs/PreviewResponseBody.tsx
@@ -1,10 +1,11 @@
 /* eslint-disable @typescript-eslint/no-unused-vars */
 import { ResponseBody } from 'src/shared/generated_models';
 import { match } from 'ts-pattern';
-import MarkdownView from '../components/response_body/MarkdownView';
+import MarkdownView from '../components/response_body/text_views/MarkdownView';
 import DirectoryView from '../components/response_body/directory_views/DirectoryView';
 import BatchDirectoryView from '../components/response_body/directory_views/BatchDirectoryView';
-import BatchFileResponseView from '../components/response_body/BatchFileResponseView';
+import BatchFileView from '../components/response_body/file_views/BatchFileView';
+import BatchTextView from '../components/response_body/text_views/BatchTextView';
 
 export default function PreviewResponseBody({
   response,
@@ -25,17 +26,13 @@ export default function PreviewResponseBody({
       return <div>Text Response</div>;
     })
     .with({ output_type: 'batchfile' }, (batchFileResponse) => {
-      return <BatchFileResponseView batchFileResponse={batchFileResponse} />;
+      return <BatchFileView response={batchFileResponse} />;
     })
     .with({ output_type: 'batchtext' }, (batchTextResponse) => {
-      return <div>Batch Text Response</div>;
+      return <BatchTextView response={batchTextResponse} />;
     })
     .with({ output_type: 'batchdirectory' }, (batchDirectoryResponse) => {
-      return (
-        <BatchDirectoryView
-          directoryResponses={batchDirectoryResponse.directories}
-        />
-      );
+      return <BatchDirectoryView response={batchDirectoryResponse} />;
     })
     .otherwise(() => {
       return <div>Unknown Response</div>;
